update file to replace some column data types from INT to BIGINT

+++ Database_Tables_SQL/stock_history.sql
@@ -7,7 +7,7 @@ CREATE TABLE stock_history (
 	low DEC,
 	close DEC,
 	adjclose DEC,
-	volume INT,
+	volume BIGINT,
 	ticker VARCHAR(20),
 	"jobTimestamp" TIMESTAMP
 );
